First attempt using dedicated .h & .cpp files
